+++ backend/SmartPlaylist/Contracts/SmartPlaylistNewItemOrderDto.cs
@@ -5,6 +5,11 @@ namespace SmartPlaylist.Contracts
     [Serializable]
     public class SmartPlaylistNewItemOrderDto
     {
+        public static SmartPlaylistNewItemOrderDto Default => new SmartPlaylistNewItemOrderDto()
+        {
+            HasSort = false,
+            OrderBy = "None"
+        };
         public bool HasSort { get; set; }
         public string OrderBy { get; set; }
     }